Addressed review comment by separating out the TransformedQueryBuilder class
