+++ document-store/src/main/java/org/hypertrace/core/documentstore/Collection.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
+import org.hypertrace.core.documentstore.query.QueryInternal;
 
 /** Interface spec for common operations on a collection of documents */
 public interface Collection {
@@ -56,8 +57,7 @@ public interface Collection {
 
   /**
    * Search for documents matching the query. For advanced functionalities like selecting summing
-   * column values, sorting by expressions, etc., ref. {@link
-   * #find(org.hypertrace.core.documentstore.query.Query)}
+   * column values, sorting by expressions, etc., ref. {@link #find(QueryInternal)}
    *
    * @param query filter to query matching documents
    * @return {@link Iterator} of matching documents
@@ -71,7 +71,7 @@ public interface Collection {
    * @param query The query definition to find
    * @return {@link Iterator} of matching documents
    */
-  Iterator<Document> find(org.hypertrace.core.documentstore.query.Query query);
+  Iterator<Document> find(final org.hypertrace.core.documentstore.query.Query query);
 
   /**
    * Aggregate the documents conforming to the query specification.
@@ -79,7 +79,7 @@ public interface Collection {
    * @param query The aggregate query specification
    * @return {@link Iterator} of matching documents
    */
-  Iterator<Document> aggregate(org.hypertrace.core.documentstore.query.Query query);
+  Iterator<Document> aggregate(final org.hypertrace.core.documentstore.query.Query query);
 
   /**
    * Delete the document with the given key.
