+++ document-store/src/main/java/org/hypertrace/core/documentstore/mongo/parser/MongoSortingExpressionParser.java
@@ -16,7 +16,7 @@ import org.hypertrace.core.documentstore.expression.impl.FunctionExpression;
 import org.hypertrace.core.documentstore.expression.impl.IdentifierExpression;
 import org.hypertrace.core.documentstore.expression.operators.SortingOrder;
 import org.hypertrace.core.documentstore.parser.SortingExpressionVisitor;
-import org.hypertrace.core.documentstore.query.Query;
+import org.hypertrace.core.documentstore.query.QueryInternal;
 import org.hypertrace.core.documentstore.query.SortingSpec;
 
 public final class MongoSortingExpressionParser implements SortingExpressionVisitor {
@@ -70,12 +70,12 @@ public final class MongoSortingExpressionParser implements SortingExpressionVisi
     return Map.of(parsed, value);
   }
 
-  public static BasicDBObject getSortClause(final Query query) {
+  public static BasicDBObject getSortClause(final QueryInternal query) {
     BasicDBObject orders = getOrders(query);
     return orders.isEmpty() ? orders : new BasicDBObject(SORT_CLAUSE, orders);
   }
 
-  public static BasicDBObject getOrders(final Query query) {
+  public static BasicDBObject getOrders(final QueryInternal query) {
     List<SortingSpec> sortingSpecs = query.getSorts();
 
     if (CollectionUtils.isEmpty(sortingSpecs)) {
