+++ document-store/src/main/java/org/hypertrace/core/documentstore/mongo/query/transformer/MongoQueryTransformer.java
@@ -2,8 +2,8 @@ package org.hypertrace.core.documentstore.mongo.query.transformer;
 
 import com.google.common.collect.ImmutableList;
 import java.util.List;
-import org.hypertrace.core.documentstore.query.Query;
-import org.hypertrace.core.documentstore.query.QueryTransformer;
+import org.hypertrace.core.documentstore.query.QueryInternal;
+import org.hypertrace.core.documentstore.query.transform.QueryTransformer;
 
 public class MongoQueryTransformer {
   // Order of transformation application matters
@@ -13,8 +13,8 @@ public class MongoQueryTransformer {
           .add(new MongoSelectionsUpdatingTransformer())
           .build();
 
-  public static Query transform(final Query query) {
-    Query transformedQuery = query;
+  public static QueryInternal transform(final QueryInternal query) {
+    QueryInternal transformedQuery = query;
 
     for (QueryTransformer transformer : TRANSFORMERS) {
       transformedQuery = transformer.transform(transformedQuery);
