+++ document-store/src/main/java/org/hypertrace/core/documentstore/mongo/query/transformer/MongoSelectionsUpdatingTransformer.java
@@ -2,14 +2,14 @@ package org.hypertrace.core.documentstore.mongo.query.transformer;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.hypertrace.core.documentstore.query.Query;
-import org.hypertrace.core.documentstore.query.Query.TransformedQueryBuilder;
-import org.hypertrace.core.documentstore.query.QueryTransformer;
+import org.hypertrace.core.documentstore.query.QueryInternal;
 import org.hypertrace.core.documentstore.query.SelectionSpec;
+import org.hypertrace.core.documentstore.query.transform.QueryTransformer;
+import org.hypertrace.core.documentstore.query.transform.TransformedQueryBuilder;
 
 final class MongoSelectionsUpdatingTransformer implements QueryTransformer {
   @Override
-  public Query transform(final Query query) {
+  public QueryInternal transform(final QueryInternal query) {
     List<SelectionSpec> newSpecs = new ArrayList<>();
 
     for (SelectionSpec spec : query.getSelections()) {
@@ -19,6 +19,6 @@ final class MongoSelectionsUpdatingTransformer implements QueryTransformer {
       newSpecs.add(newSpec);
     }
 
-    return new TransformedQueryBuilder(query).buildWithSelections(newSpecs);
+    return (QueryInternal) new TransformedQueryBuilder(query).replaceSelections(newSpecs).build();
   }
 }
