--- document-store/src/main/java/org/hypertrace/core/documentstore/query/QueryTransformer.java
@@ -1,5 +1,0 @@
-package org.hypertrace.core.documentstore.query;
-
-public interface QueryTransformer {
-  Query transform(final Query query);
-}
